requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
catboost>=1.2
shap>=0.45
SQLAlchemy>=2.0
PyMySQL>=1.1.0
python-dotenv>=1.0
joblib>=1.3
tqdm>=4.66
flask>=2.3
gunicorn>=21.0
